backend/requirements.txt
pandas==2.1.4
python-docx==1.1.0
python-dotenv==1.0.0
msal==1.26.0
requests==2.31.0
openpyxl==3.1.2